LayoutTests/indieui/uiactions.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="item" uiactions="valuechange redo">a</div>

<div id="console"></div>

<script>
    description("This tests that we can get and set uiActions.");

    var item = document.getElementById("item");
    shouldBe("item.uiactions", "'valuechange redo'");
    item.uiactions = "undo";
    shouldBe("item.uiactions", "'undo'");

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
