safari-600.7-branch/LayoutTests/indieui/uiactions-expected.txt

a
This tests that we can get and set uiActions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item.uiactions is 'valuechange redo'
PASS item.uiactions is 'undo'
PASS successfullyParsed is true

TEST COMPLETE
